Move confluent_kafka_output to connector
