+++ tests/unit/connector/test_connector_factory.py
@@ -20,7 +20,7 @@ from logprep.connector.confluent_kafka_input import ConfluentKafkaInput
 from logprep.input.dummy_input import DummyInput
 from logprep.input.json_input import JsonInput
 from logprep.input.jsonl_input import JsonlInput
-from logprep.output.confluent_kafka_output import ConfluentKafkaOutput
+from logprep.connector.confluent_kafka_output import ConfluentKafkaOutput
 from logprep.output.dummy_output import DummyOutput
 from logprep.output.es_output import ElasticsearchOutput
 from logprep.output.writing_output import WritingOutput
